Add support for Argon2d/Argon2id variants
This is a backwards-compatible change to enable usage of the
Argon2d and Argon2id hashes. Everything defaults to use of Argon2i (or
autodetection, for the argon2-verify routine), so it should remain
compatible with prior uses. However, various sources recommend the
argon2_id variant by default, since it provides a balance of protection
against side-channel attacks and GPU attacks.

Resolves skinkade/p6-crypt-argon2#5.